Moved dynamic tension and compositeLine functions to existing files
- functions from dynamic_tension_functions.py are now on helpers
- Because the dynamic calculations depend on the number of nodes in the subsystem/line, I slightly modified functions lines2subsystem and makeGeneric to specify the number of segments to be used in addLine.
- I have not used subsystem.makeFromSystem or getNodes yet
