+++ moorpy/subsystem.py
@@ -119,7 +119,7 @@ class Subsystem(System, Line):
         self.qs = 1         # flag that it's a MoorPy analysis, so System methods don't complain. One day should replace this <<<
     
     
-    def makeGeneric(self, lengths, types, suspended=0):
+    def makeGeneric(self, lengths, types, suspended=0, nSegs=40):
         '''Creates a cable of n components going between an anchor point and
         a floating body (or a bridle point). If shared, it goes between two
         floating bodies.
@@ -184,8 +184,16 @@ class Subsystem(System, Line):
             else:
                 raise Exception(f"Can't find lineType '{types[i]}' in the SubSystem.")
             
+            # add the line segment using the reference to its lineType dict            
             # add the line segment using the reference to its lineType dict
-            self.addLine(lengths[i], self.lineTypes[i])
+            if nSegs is None:
+                self.addLine(lengths[i], self.lineTypes[i])
+            elif isinstance(nSegs, (int, float)):
+                self.addLine(lengths[i], self.lineTypes[i], nSegs=nSegs)
+            elif isinstance(nSegs, list):
+                self.addLine(lengths[i], self.lineTypes[i], nSegs=nSegs[i])
+            else:
+                raise ValueError("Invalid type for nSegs. Expected None, a number, or a list.")
 
             # add the upper end point of the segment
             if i==self.nLines-1:                            # if this is the upper-most line
@@ -239,7 +247,7 @@ class Subsystem(System, Line):
             line_id = point.attached[-1] # get next line's id
             line = sys.lineList[line_id - 1] # get line object
             self.lineTypes[line.type['name']] = dict(line.type)  # copy the lineType dict
-            self.addLine(line.L0, self.lineTypes[line.type['name']]) # add the line
+            self.addLine(line.L0, self.lineTypes[line.type['name']], nSegs=line.nNodes-1) # add the line
             
 
             # get the next point
@@ -249,7 +257,7 @@ class Subsystem(System, Line):
             pointB_id = attached_points[0] # get second point id
             point = sys.pointList[pointB_id-1] # get second point object
 
-            if line(point.attached) == 1:  # must be the endpoint
+            if len(point.attached) == 1:  # must be the endpoint
                 self.addPoint(-1, point.r)
             else:  # intermediate point along line
                 self.addPoint(0, point.r, DOFs=[0,2]) # may need to ensure there's no y component
